requirements.txt
streamlit==1.32.2
streamlit-cropper==0.2.2
Pillow==10.3.0
rembg==2.0.57
opencv-python==4.9.0.80
numpy==1.26.4
